[REF] Adds images and transforms modules
