+++ abagen/correct.py
@@ -12,7 +12,7 @@ import pandas as pd
 import scipy.stats as sstats
 from scipy.spatial.distance import cdist
 
-from . import utils
+from . import images, utils
 
 
 def _unpack_tuple(var):
@@ -453,7 +453,7 @@ def remove_distance(coexpression, atlas, atlas_info=None, labels=None):
     """
 
     # load atlas_info, if provided
-    atlas = utils.check_img(atlas)
+    atlas = images.check_img(atlas)
     if atlas_info is not None:
         atlas_info = utils.check_atlas_info(atlas, atlas_info, labels=labels)
         if labels is not None and len(labels) != len(coexpression):
@@ -468,7 +468,7 @@ def remove_distance(coexpression, atlas, atlas_info=None, labels=None):
 
     # we'll do basic Euclidean distance correction for now
     # TODO: implement gray matter volume / cortical surface path distance
-    centroids = utils.get_centroids(atlas, labels=labels)
+    centroids = images.get_centroids(atlas, labels=labels)
     dist = cdist(centroids, centroids, metric='euclidean')
 
     corr_resid = np.zeros_like(coexpression)
